pyproject.toml
[project]
name = "tgc-mvp"
version = "0.1.0"
description = "TGC RAG MVP: vector storage and chatbot over Gospel Coalition articles"
requires-python = ">=3.10"
readme = "README.md"
license = { text = "Proprietary" }

dependencies = [
  "requests>=2.31.0",
  "beautifulsoup4>=4.12.0",
  "lxml>=4.9.0",
  "chromadb>=0.4.0",
  "openai>=1.0.0",
  "python-dotenv>=1.0.0",
  "tenacity>=8.2.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
  "pytest-cov>=4.0.0",
  "ruff>=0.1.0",
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.ruff]
line-length = 100
target-version = "py310"
